% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{all.genes.lengths}
\alias{all.genes.lengths}
\title{Coding genes length by Ensamble, length is calculated as sum of all coding exons (merged before so each position only once ocunted).}
\format{data frame}
\source{
http://www.genenames.org/ (Most recent HUGO approved names ) and ftp://ftp.ensembl.org/pub/current_gtf/homo_sapiens (info for genes, exons, lenghts)
}
\usage{
all.genes.lengths
}
\description{
This data frame containts information about genes length.
There is 19202 rows and each row represents gene. This set of genes is acquired from HGNC website and only approved and protein coding genes are included..
Rownames are Hugo symbols for genes.
Data frame has 2 columns.
}
\section{Variables}{

\itemize{
 \item Hugo_Symbol same as rowname. Hugo symbol as unique indentifier for genes.
 \item Length length of genes reported in NCBI
 }
}
\keyword{data}

