% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bayes.R
\name{bayes.driver}
\alias{bayes.driver}
\title{Bayesina risk inference model.}
\usage{
bayes.driver(sample.mutations, bcgr.prob, genes = NULL,
  prior.driver = 127/20000, gene.mut.driver = NULL, driver.genes = NULL,
  Variant_Classification = NULL, Hugo_Symbol = NULL,
  Tumor_Sample_Barcode = NULL, CCF = NULL, Damage_score = NULL,
  mode = "MAX", epsilon = 0.05)
}
\arguments{
\item{sample.mutations}{data frame with SNVs and InDels in MAF like format.
Columns (with exactly same names) which \code{sample.mutations} should have are:
\itemize{
     \item Variant_Classification column specifed by MAF format, used to distinguish between silent and nonsilent SNVs
     \item Hugo_Symbol column specifed by MAF format, which reports gene for each SNV.
     \item Tumor_Sample_Barcode column specifed by MAF format, reporting for each SNV in wich patient was found.
     \item CCF numeric column produce by \code{CCF} function.
     \item Damage_score numeric column with values between 0 and 1, where 1 means very damaging SNV/IndDel and 0 not damaging SNV/InDel
}}

\item{bcgr.prob}{a numeric vector, same lenght as genes (should be same orderd also) which gives probability of gene having somatic mutation in healfy population.
There are functions for obtaining this vector: \code{bcgr}, \code{bcgr.lawrence} and \code{bcgr.combine}.}

\item{genes}{a vector of genes which were sequenced.
They should be unique values of Hugo_Symbol column (with possibility of more additional genes which did not have any SNV/Indel. in given cohort). Default NULL.}

\item{prior.driver}{a numeric value representing prior probability that random gene is dirver.
Default is set to 127/20000, as it has been reported around 127 genes from ~20000 that have been involved in causing tumor.}

\item{gene.mut.driver}{a numeric value or named vector representing likelihood that gene is mutated if it is knowen to be driver.
Gene does not need to be mutated if it is driver, as cancers are heterogenious. Default is set to NULL and driver.genes are considered as drivers.}

\item{driver.genes}{a character vector of genes which are considered as drivers for this cancer. If NULL then used set is \code{driver.genes.concensus}.}

\item{Variant_Classification}{(optional) integer/numeric value indicating column in \code{sample.mutations} which contain classification for SNV (Silent or not).
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Hugo_Symbol}{(optional) integer/numeric value indicating column in \code{sample.mutations} having gene names for reported SNVs/Indels.
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Tumor_Sample_Barcode}{(optional) integer/numeric value indicating column in \code{sample.mutations} which have sample ids for SNVs/Indels.
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{CCF}{(optional) integer/numeric value indicating column in \code{sample.mutations} which have cancer cell fraction information for SNVs/Indels.
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Damage_score}{(optional) integer/numeric value indicating column in \code{sample.mutations} which contain damage score for SNVs/Indels.
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{mode}{a charechter value indicationg how to solve when in one gene-sample pair there are multiple mutations. Options are SUM, MAX and ADVANCE}

\item{epsilon}{a numeric value. If mode is ADVANCE, epsilone value will be threshold for CCF difference to decide if they are in same or different clone.}
}
\value{
a data frame with ranked genes by posteriory probability of gene beeing true driver.
Additional columns with usefull info are contained in data frame.
}
\description{
\code{bayes.driver} function performs by runing Bayesian driver inference model where each observed mutation in gene is taken as proof for being true driver.
}
\examples{
\donttest{
# first calculate CCF
sample.genes.mutect <- CCF(sample.genes.mutect)
# then somatic background probability
bcgr.prob <- bcgr.combine(sample.genes.mutect)
# bayes risk model
driver.genes <- bayes.driver(sample.genes.mutect, driver = 0.001, gene.mut.driver=1/50)
head(driver.genes)
}
}
\seealso{
\code{\link{bcgr}}, \code{\link{bcgr.lawrence}}  and \code{\link{bcgr.combine}} for obtaining bcgr.prob variable.
}
\keyword{risk}

