% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{bayes.risk}
\alias{bayes.risk}
\title{Bayesina risk inference model.}
\usage{
bayes.risk(sample.mutations, bcgr.prob, genes = NULL, prior.sick = 0.0045,
  Variant_Classification = NULL, Hugo_Symbol = NULL,
  Tumor_Sample_Barcode = NULL, CCF = NULL, Damage_score = NULL,
  mode = "MAX", epsilon = 0.05)
}
\arguments{
\item{sample.mutations}{data frame with SNVs and InDels in MAF like format.
Columns (with exactly same names) which \code{sample.mutations} should have are: 
\itemize{ 
     \item Variant_Classification column specifed by MAF format, used to distinguish between silent and nonsilent SNVs
     \item Hugo_Symbol column specifed by MAF format, which reports gene for each SNV.
     \item Tumor_Sample_Barcode column specifed by MAF format, reporting for each SNV in wich patient was found. 
     \item CCF numeric column produce by \code{CCF} function.
     \item Damage_score numeric column with values between 0 and 1, where 1 means very damaging SNV/IndDel and 0 not damaging SNV/InDel
}}

\item{bcgr.prob}{a numeric vector, same lenght as genes (should be same orderd also) which gives probability of gene having somatic mutation in healfy population.
There are functions for obtaining this vector: \code{bcgr}, \code{bcgr.lawrence} and \code{bcgr.combine}.}

\item{genes}{vector of genes which were sequenced. 
They should be unique values of Hugo_Symbol column (with possibility of more additional genes which did not have any SNV/Indel. in given cohort). Default NULL.}

\item{prior.sick}{a numeric value representing incidence of tumor in population. Set by default to 0.0045 .}

\item{Variant_Classification}{(optional) integer/numeric value indicating column in \code{sample.mutations} which contain classification for SNV (Silent or not). 
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Hugo_Symbol}{(optional) integer/numeric value indicating column in \code{sample.mutations} having gene names for reported SNVs/Indels.
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Tumor_Sample_Barcode}{(optional) integer/numeric value indicating column in \code{sample.mutations} which have sample ids for SNVs/Indels. 
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{CCF}{(optional) integer/numeric value indicating column in \code{sample.mutations} which have cancer cell fraction information for SNVs/Indels. 
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Damage_score}{(optional) integer/numeric value indicating column in \code{sample.mutations} which contain damage score for SNVs/Indels. 
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{mode}{a charechter value indicationg how to solve when in one gene-sample pair there are multiple mutations. Options are SUM, MAX and ADVANCE}

\item{epsilon}{a numeric value. If mode is ADVANCE, epsilone value will be threshold for CCF difference to decide if they are in same or different clone.}
}
\value{
a data frame with ranked genes by posteriory probability of gene beeing risk factor for developing tumor. 
Additional columns with usefull info are contained in data frame.
}
\description{
\code{bayes.risk} function performs by runing Bayesian risk inference model when priors are set by user and liklihood is calculated from given data of SNVs/InDels.
}
\examples{
\donttest{
# first calculate CCF
sample.genes.mutect <- CCF(sample.genes.mutect)
# then somatic background probability
bcgr.prob <- bcgr.combine(sample.genes.mutect)
# bayes risk model
risk.genes <- bayes.risk(sample.genes.mutect,  bcgr.prob, prior.sick = 0.00007) 
head(risk.genes)  
}
}
\seealso{
\code{\link{bcgr}}, \code{\link{bcgr.lawrence}}  and \code{\link{bcgr.combine}} for obtaining bcgr.prob variable.
}
\keyword{risk}

