% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{boxplotCCF.mutations}
\alias{boxplotCCF.mutations}
\title{Plot boxplot of mutations for top genes.}
\usage{
boxplotCCF.mutations(sample.mutations, result.df, topGenes = 20,
  silent = FALSE, indels = TRUE, Tumor_Sample_Barcode = NULL,
  Hugo_Symbol = NULL, CCF = NULL, Variant_Classification = NULL,
  Variant_Type = NULL, color = NULL, shape = NULL)
}
\arguments{
\item{sample.mutations}{data frame in MAF like format.
Columns (with exactly same names) which \code{sample.mutations} should have are: 
\itemize{ 
     \item Tumor_Sample_Barcode column specifed by MAF format, reporting for each SNV in wich patient was found. 
     \item Hugo_Symbol column specifed by MAF format, which reports gene for each SNV.
     \item CCF numeric column produce by \code{CCF} function.
     \item Variant_Classification column specifed by MAF format, used to distinguish between silent and nonsilent SNVs
     \item Variant_Type columns pecifed by MAF format; is mutation SNV or InDel
}}

\item{result.df}{a data frame with Hugo_Symbol column. Genes in this column should be ordered by importance.}

\item{topGenes}{a numeric/integer value; How meny top genes from results will be ploted.}

\item{silent}{a boolean value indicating should silent mutations be counted. By default it is True.}

\item{indels}{a boolean value indicating should indels be counted. By default it is True.}

\item{Tumor_Sample_Barcode}{(optional) integer/numeric value indicating column in \code{sample.mutations} which have sample ids for SNVs/Indels. 
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Hugo_Symbol}{(optional) integer/numeric value indicating column in \code{sample.mutations} having gene names for reported SNVs/Indels.
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{CCF}{(optional) integer/numeric value indicating column in \code{sample.mutations} which have cancer cell fraction information for SNVs/Indels. 
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Variant_Classification}{(optional) integer/numeric value indicating column in \code{sample.mutations} which contain classification for SNV (Silent or not). 
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Variant_Type}{(optional) integer/numeric value indicating column in \code{sample.mutations} which contains indormation if mutations is SNV or InDel .
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{color}{a numeric or character value indicating column to color the samples.}

\item{shape}{a numeric or character value indicating column to use as shape for the samples.}
}
\value{
ggplot2 object
}
\description{
\code{boxplotCCF.mutations} Function to plot CCF's of all mutations for top genes.
}
\examples{
\donttest{
#get CCF column
sample.genes.mutect <- CCF(sample.genes.mutect)
# get background
bcgr.prob <- bcgr.combine(sample.genes.mutect, length.genes$Hugo_Symbol, length.genes$Coverd_len)
# get ranking
df1 <- bayes.risk(sample.genes.mutect, bcgr.prob, prior.sick = 0.00007) 
df2 <- bayes.driver(sample.genes.mutect, bcgr.prob,  prior.driver = 0.001) 
df.final <- combine.ranking(list(df1, df2),  min.mut = 2 )
# plot boxplot of CCF for top genes
boxplotCCF(sample.mutations=sample.genes.mutect, result.df=df.final)
}
}

