% Generated by roxygen2 (4.0.1): do not edit by hand
\name{boxplotCCF.patients}
\alias{boxplotCCF.patients}
\title{Plot boxplot of patients mutations CCF's for top genes.}
\usage{
boxplotCCF.patients(sample.mutations, result.df, topGenes = 20,
  silent = FALSE, indels = TRUE, mode = "MAX", epsilon = 0.05,
  sample.gene.lim = 1, Tumor_Sample_Barcode = NULL, Hugo_Symbol = NULL,
  CCF = NULL, Variant_Classification = NULL, Variant_Type = NULL,
  color = NULL, shape = NULL)
}
\arguments{
\item{sample.mutations}{data frame in MAF like format.
Columns (with exactly same names) which \code{sample.mutations} should have are:
\itemize{
     \item Tumor_Sample_Barcode column specifed by MAF format, reporting for each SNV in wich patient was found.
     \item Hugo_Symbol column specifed by MAF format, which reports gene for each SNV.
     \item CCF numeric column produce by \code{CCF} function.
     \item Variant_Classification column specifed by MAF format, used to distinguish between silent and nonsilent SNVs
     \item Variant_Type columns pecifed by MAF format; is mutation SNV or InDel
}}

\item{result.df}{a data frame with Hugo_Symbol column. Genes in this column should be ordered by importance.}

\item{topGenes}{a numeric/integer value; How meny top genes from results will be ploted.}

\item{silent}{a boolean value indicating should silent mutations be counted. By default it is True.}

\item{indels}{a boolean value indicating should indels be counted. By default it is True.}

\item{mode}{a charechter value indicationg how to solve when in one gene-sample pair there are multiple mutations. Options are SUM, MAX and ADVANCE}

\item{epsilon}{a numeric value. If mode is ADVANCE, epsilone value will be threshold for CCF difference to decide if they are in same or different clone.}

\item{sample.gene.lim}{a numeric value specifying upper limit when summing is perfomed for gene-sample pair}

\item{Tumor_Sample_Barcode}{(optional) integer/numeric value indicating column in \code{sample.mutations} which have sample ids for SNVs/Indels.
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Hugo_Symbol}{(optional) integer/numeric value indicating column in \code{sample.mutations} having gene names for reported SNVs/Indels.
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{CCF}{(optional) integer/numeric value indicating column in \code{sample.mutations} which have cancer cell fraction information for SNVs/Indels.
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Variant_Classification}{(optional) integer/numeric value indicating column in \code{sample.mutations} which contain classification for SNV (Silent or not).
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Variant_Type}{(optional) integer/numeric value indicating column in \code{sample.mutations} which contains indormation if mutations is SNV or InDel .
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{color}{(optional) charachter value indicating column in \code{sample.mutations} to used as color dots.
Default is NULL value and grey dots gonna be ploted. This variable should be unique by patients.}

\item{shape}{(optional) charachter value indicating column in \code{sample.mutations} to used as shape for dots .
Default is NULL value and grey dots gonna be ploted. This variable should be unique by patients.}
}
\value{
ggplot2 object
}
\description{
\code{boxplotCCF.patients} Function to plot CCF's of  aggregated patient-gene mutatuons, for top genes.
}
\examples{
\donttest{
#get CCF column
sample.genes.mutect <- CCF(sample.genes.mutect)
# get background
bcgr.prob <- bcgr.combine(sample.genes.mutect, length.genes$Hugo_Symbol, length.genes$Coverd_len)
# get ranking
df1 <- bayes.risk(sample.genes.mutect, bcgr.prob, genes=length.genes$Hugo_Symbol, prior.sick = 0.00007)
df2 <- bayes.driver(sample.genes.mutect, bcgr.prob, length.genes$Hugo_Symbol, prior.driver = 0.001)
df.final <- combine.ranking(list(df1, df2), length.genes$Hugo_Symbol, min.mut = 2 )
# plot boxplot of CCF for top genes
boxplotCCF(sample.mutations=sample.genes.mutect, result.df=df.final)
}
}

