% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cDriver-package.r
\docType{package}
\name{cDriver}
\alias{cDriver}
\alias{cDriver-package}
\title{cDriver.}
\description{
\code{cDriver} R package for finding candidate driver genes in cancers. 
Useful functions:
\itemize{ 
     \item \code{CCF}
     \item \code{bcgr.lawrence}
     \item \code{bcgr}
     \item \code{bcgr.combine}
     \item \code{bayes.risk}
     \item \code{bayes.driver}
     \item \code{combine.ranking}
     \item \code{cut.off.risk}
     \item \code{plotSamplesMut}
     \item \code{plotMutChange}
     \item \code{boxplotCCF.mutations}
     \item \code{boxplotCCF.patients}
     \item \code{plotStaircase}
     \item \code{boxplotCCF.mutations} }
Useful data:
\itemize{ 
     \item \code{lawrence.df}
     \item \code{length.genes}
     \item \code{driver.genes.concensus} 
     \item \code{sample.genes.mutect}}
}

