% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{combine.ranking}
\alias{combine.ranking}
\title{Combining mulitple methods for ranking genes.}
\usage{
combine.ranking(rankings = list(bayes.risk.rank, bayes.driver.rank),
  genes = NULL, min.mut = 2)
}
\arguments{
\item{rankings}{a list containing at least two elements, first one data frame output of bayes.risk and second  data frame output of bayes.driver functions.
All next elements of the list, should be named numeric vectors, where nemas are Hugo_Symbol and value represent any score where higer values are scoring better for gene to be driver.}

\item{genes}{a vector of genes which were sequenced/analysed. 
They should be unique values of Hugo_Symbol column (with possibility of more additional genes which did not have any SNV/Indel. in given cohort).}

\item{min.mut}{a numeric value, threshold for filtering genes based on minimum number of nonsilent mutations.}
}
\value{
a data frame with ranked genes by combining different methods ranking. 
Additional columns with usefull info are contained in resulting data frame also.
}
\description{
\code{combine.ranking} function combines mulitple rankings to one final. All the rankings takes the same importans for final decision.
}
\seealso{
\code{\link{bayes.risk}}, \code{\link{bayes.driver}}   for obtaining bayes.risk.rank and bayes.driver.rank variables.
}
\keyword{combine}
