% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{driver.genes.concensus}
\alias{driver.genes.concensus}
\title{Vector of genes which are considered as common cancer drivers.}
\format{charachter vector}
\usage{
all.genes.lengths
}
\description{
This charachter vector containes 544 cammon cancer genes.
}
\keyword{data}
