% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{lawrence.df}
\alias{lawrence.df}
\title{Lawrence's paper information}
\format{data frame}
\source{
Suplementary matirial from Lawrence paper: \url{http://www.ncbi.nlm.nih.gov/pubmed/23770567}.
}
\usage{
lawrence.df
}
\description{
This data frame containts information used in package which is provided in Lawrence paper.
There is 18859 rows and each row represents gene. Rownames are Hugo symbols for genes. They were corrected to most recent used Hugo gene name.
Data frame has 4 columns.
}
\details{
Data that are included in cDriver package
}
\section{Variables}{

\itemize{
 \item N_nonsilent number of all possible nonsilent substitutions
 \item N_silent number of all possible silent substitutions
 \item N_noncoding number of all possible noncoding substitutions
 \item noncoding_mutation_rate measured noncoding mutation rate - somatic background mutation rate
 }
}
\keyword{data}

