% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{length.genes}
\alias{length.genes}
\title{Genes length by NCBI, length which was covered by sequencing,}
\format{data frame}
\usage{
length.genes
}
\description{
This data frame containts information about genes length. 
There is 16485 rows and each row represents gene. This set of genes were analysed as part of CLL somatic driver genes at CRG and ICGC. 
Rownames are Hugo symbols for genes. They were corrected to most recent used Hugo gene name.
Data frame has 5 columns.
}
\section{Variables}{

\itemize{
 \item Hugo_Symbol same as rowname. Hugo symbol as unique indentifier for genes.
 \item Coverd_len length of the gene which was coverd by sequncing CLL project.
 \item Length length of genes reported in NCBI
 \item procentege which was covered of total gene (Coverd_len / Length)
 \item p_by_len  probability that gene get one mutation, if each base pair had same chance; Coverd_len/sum(length.genes$Coverd_len)
 }
}

\keyword{data}
