% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotMutChange}
\alias{plotMutChange}
\title{Plot mutations pattern frequncies.}
\usage{
plotMutChange(sample.mutations, silent = TRUE, fill = TRUE,
  Tumor_Sample_Barcode = NULL, Variant_Type = NULL,
  Variant_Classification = NULL, Reference_Allele = NULL,
  Tumor_Seq_Allele2 = NULL)
}
\arguments{
\item{sample.mutations}{data frame in MAF like format.
Columns (with exactly same names) which \code{sample.mutations} should have are:
\itemize{
     \item Tumor_Sample_Barcode column specifed by MAF format, reporting for each SNV in wich patient was found.
     \item Variant_Type columns pecifed by MAF format; is mutation SNV or InDel
     \item Variant_Classification column specifed by MAF format, used to distinguish between silent and nonsilent SNVs
     \item Reference_Allele column specifed by MAF format, represents referent allele
     \item Tumor_Seq_Allele2 column specifed by MAF format, represents alternative allele
}}

\item{silent}{a boolean value indicating should silent mutations be counted. By default it is True.}

\item{fill}{a boolean value indicating should plot only represent proportion between 3 types of mutations, not counts. By default it is True}

\item{Tumor_Sample_Barcode}{(optional) integer/numeric value indicating column in \code{sample.mutations} which have sample ids for SNVs/Indels.
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Variant_Type}{(optional) integer/numeric value indicating column in \code{sample.mutations} which contains indormation if mutations is SNV or InDel .
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Variant_Classification}{(optional) integer/numeric value indicating column in \code{sample.mutations} which contain classification for SNV (Silent or not).
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Reference_Allele}{(optional) integer/numeric value indicating column in \code{sample.mutations} which contain reference alleles.
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Tumor_Seq_Allele2}{(optional) integer/numeric value indicating column in \code{sample.mutations} which contain alternative alleles.
Default is NULL value (in this case \code{sample.mutations} should already have this column)}
}
\value{
ggplot2 object
}
\description{
\code{plotMutChange} Function to plot frequency of mutation changes.
}
\examples{
\donttest{
# plot SNVs changes patterns
plotMutChange(sample.genes.mutect)
# plot SNVs changes patterns only for non silent mutaions
plotMutChange(sample.genes.mutect, silent=FALSE, fill=FALSE)
}
}

