% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSamplesMut}
\alias{plotSamplesMut}
\title{Plot mutations per sample distribution.}
\usage{
plotSamplesMut(sample.mutations, indels = TRUE, silent = TRUE,
  fill = FALSE, Tumor_Sample_Barcode = NULL, Hugo_Symbol = NULL,
  Variant_Classification = NULL, Variant_Type = NULL)
}
\arguments{
\item{sample.mutations}{data frame in MAF like format.
Columns (with exactly same names) which \code{sample.mutations} should have are: 
\itemize{ 
     \item Variant_Classification column specifed by MAF format, used to distinguish between silent and nonsilent SNVs
     \item Hugo_Symbol column specifed by MAF format, which reports gene for each SNV.
     \item Tumor_Sample_Barcode column specifed by MAF format, reporting for each SNV in wich patient was found. 
     \item Variant_Type columns pecifed by MAF format; is mutation SNV or InDel
}}

\item{indels}{a boolean value indicating should indels be counted. By default it is True.}

\item{silent}{a boolean value indicating should silent mutations be counted. By default it is True.}

\item{fill}{a boolean value indicating should plot only represent proportion between 3 types of mutations, not counts. By default it is False.}

\item{Tumor_Sample_Barcode}{(optional) integer/numeric value indicating column in \code{sample.mutations} which have sample ids for SNVs/Indels. 
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Hugo_Symbol}{(optional) integer/numeric value indicating column in \code{sample.mutations} having gene names for reported SNVs/Indels.
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Variant_Classification}{(optional) integer/numeric value indicating column in \code{sample.mutations} which contain classification for SNV (Silent or not). 
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Variant_Type}{(optional) integer/numeric value indicating column in \code{sample.mutations} which contains indormation if mutations is SNV or InDel .
Default is NULL value (in this case \code{sample.mutations} should already have this column)}
}
\value{
ggplot2 object
}
\description{
\code{plot.samplesMut} Function to plot samples mutation counts.
}
\examples{
\donttest{
# plot sample's mutations , all of them
plotSamplesMut(sample.genes.mutect)
# plot proportion of silent and nonsilent, without indels
plotSamplesMut(sample.genes.mutect, indels=FALSE, fill=TRUE)
}
}

