% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotStaircase}
\alias{plotStaircase}
\title{Plot heatmap for genes-sample pairs .}
\usage{
plotStaircase(sample.mutations, result.df, topGenes = 40, silent = FALSE,
  indels = TRUE, allSamples = FALSE, Tumor_Sample_Barcode = NULL,
  Hugo_Symbol = NULL, Variant_Classification = NULL, Variant_Type = NULL,
  color = "CCF", groupMax = 1, order = "frequency")
}
\arguments{
\item{sample.mutations}{data frame in MAF like format.
Columns (with exactly same names) which \code{sample.mutations} should have are:
\itemize{
     \item Tumor_Sample_Barcode column specifed by MAF format, reporting for each SNV in wich patient was found.
     \item Hugo_Symbol column specifed by MAF format, which reports gene for each SNV.
     \item Variant_Classification column specifed by MAF format, used to distinguish between silent and nonsilent SNVs
     \item Variant_Type columns pecifed by MAF format; is mutation SNV or InDel
}}

\item{result.df}{a data frame with Hugo_Symbol column. Genes in this column should be ordered by importance.}

\item{topGenes}{a numeric/integer value; How meny top genes from results will be ploted.}

\item{silent}{a boolean value indicating should silent mutations be counted. By default it is True.}

\item{indels}{a boolean value indicating should indels be counted. By default it is True.}

\item{allSamples}{a boolean value indicating if all samples should be plotted.}

\item{Tumor_Sample_Barcode}{(optional) integer/numeric value indicating column in \code{sample.mutations} which have sample ids for SNVs/Indels.
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Hugo_Symbol}{(optional) integer/numeric value indicating column in \code{sample.mutations} having gene names for reported SNVs/Indels.
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Variant_Classification}{(optional) integer/numeric value indicating column in \code{sample.mutations} which contain classification for SNV (Silent or not).
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Variant_Type}{(optional) integer/numeric value indicating column in \code{sample.mutations} which contains indormation if mutations is SNV or InDel .
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{color}{a numeric or character value indicating column to color the samples.}

\item{groupMax}{a numeric value which will be taken as maximum value when grouping color column for samples and genes pairs.}

\item{order}{a character value which can be or \code{'frequency'} or \code{'CCF_sum'}.}
}
\value{
ggplot2 object
}
\description{
\code{plotStaircase} Function to plot heatmap for genes-sample pairs orderd by genes which explain most of the patients .
}
\examples{
\donttest{
#get CCF column
sample.genes.mutect <- CCF(sample.genes.mutect)
# get background
bcgr.prob <- bcgr.combine(sample.genes.mutect)
# get ranking
df1 <- bayes.risk(sample.genes.mutect, bcgr.prob, prior.sick = 0.00007)
df2 <- bayes.driver(sample.genes.mutect, bcgr.prob, prior.driver = 0.001)
df.final <- combine.ranking(list(df1, df2), min.mut = 2 )
# plot boxplot of CCF for top genes
plotStaircase(sample.mutations=sample.genes.mutect, result.df=df.final)
}
}

