% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{sample.genes.mutect}
\alias{sample.genes.mutect}
\title{Genes length by NCBI, length which was covered by sequencing,}
\format{data grame}
\usage{
sample.genes.mutect
}
\description{
This data frame containts information about SNVs and InDels that are reported as part of CLL project.
There is 7860 rows and each row represents annotated SNV/InDel reported in one of 337 patients with CLL.
These are filtered SNVs and InDels for segmental duplication, exonix or splicing, intersection of used kits region, etc.
They follow MAF like format with additional columns like ploidy, CCF_CNV, etc.
Data frame has 23 columns.
}
\section{Variables}{

\itemize{
 \item Chromosome
 \item Start_Position
 \item Reference_Allele
 \item Tumor_Seq_Allele2
 \item Hugo_Symbol
 \item Tumor_Sample_Barcode
 \item VAF
 \item Function.Refseq
 \item ExonicFunction.Refseq
 \item ...
 \item ploidy
 \item CCF_CNV
 \item Damege_score
 \item gender
 \item Variant_Classification
 \item Variant_Type
 }
}
\keyword{data}

