% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bcgr.combine}
\alias{bcgr.combine}
\title{Combining two somatic background mutation probability, outputs of functions bcgr.lawrence and bcgr}
\usage{
bcgr.combine(sample.mutations, genes = NULL, lengthGenes = NULL,
  Variant_Classification = NULL, Hugo_Symbol = NULL,
  Tumor_Sample_Barcode = NULL, CCF = NULL)
}
\arguments{
\item{sample.mutations}{data frame in MAF like format.
Columns (with exactly same names) which \code{sample.mutations} should have are:
\itemize{
     \item Variant_Classification column specifed by MAF format, used to distinguish between silent and nonsilent SNVs
     \item Hugo_Symbol column specifed by MAF format, which reports gene for each SNV.
     \item Tumor_Sample_Barcode column specifed by MAF format, reporting for each SNV in wich patient was found.
     \item CCF numeric column produce by \code{CCF} function.
}}

\item{genes}{vector of genes which were sequenced.
They should be unique values of Hugo_Symbol column (with possibility of more additional genes which did not have any SNV in given cohort).
Default is NULL value and then list of unique genes is takend from \code{sample.mutations}.}

\item{lengthGenes}{numeric vector of lengths (suquenced) for \code{genes}. Should be given in same order as variable genes.
#' Default is NULL value and then Length of genes is taken from data set \code{length.genes} (form package \code{cDriver}) as column Length.
If gene is not found in this data frame, then median value is taken from listed genes in this data frame.}

\item{Variant_Classification}{(optional) integer/numeric value indicating column in \code{sample.mutations} which contain classification for SNV (Silent or not).
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Hugo_Symbol}{(optional) integer/numeric value indicating column in \code{sample.mutations} having gene names for reported SNVs.
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Tumor_Sample_Barcode}{(optional) integer/numeric value indicating column in \code{sample.mutations} which have sample ids for SNVs.
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{CCF}{(optional) integer/numeric value indicating column in \code{sample.mutations} which have cancer cell fraction information for SNVs.
Default is NULL value (in this case \code{sample.mutations} should already have this column)}
}
\value{
a named numeric vector of probabilites that gene has nonsilent mutation (not caused by cancer).
}
\description{
\code{bcgr.combine} function calculates both somatic background mutation probabilities (using bcgr.lawrence and bcgr functions) and take average value for each gene
}
\examples{
# First we need to calculate CCF
sample.genes.mutect <- CCF(sample.genes.mutect)
# Calculate somatic nonsilent background mutation probability
 background <- bcgr.combine(sample.genes.mutect, length.genes$Hugo_Symbol, length.genes$Coverd_len)
 head(background)
}
\references{
\url{http://www.ncbi.nlm.nih.gov/pubmed/23770567}.
}
\keyword{Lawrence}

