% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bcgr}
\alias{bcgr}
\title{Background mutation rate calculated based on silent mutations.}
\usage{
bcgr(sample.mutations, genes = NULL, Variant_Classification = NULL,
  Hugo_Symbol = NULL, Tumor_Sample_Barcode = NULL, CCF = NULL)
}
\arguments{
\item{sample.mutations}{data frame in MAF like format with nonsilent and silent mutations.
Columns (with exactly same names) which \code{sample.mutations} should have are:
\itemize{
     \item Variant_Classification column specifed by MAF format, used to distinguish between silent and nonsilent SNVs
     \item Hugo_Symbol column specifed by MAF format, which reports gene for each SNV.
     \item Tumor_Sample_Barcode column specifed by MAF format, reporting for each SNV in wich patient was found.
     \item CCF numeric column produce by \code{CCF} function.
}}

\item{genes}{vector of genes which were sequenced.
They should be unique values of Hugo_Symbol column (with possibility of more additional genes which did not have any SNV in given cohort).
Default is NULL value and then list of unique genes is takend from \code{sample.mutations}.}

\item{Variant_Classification}{(optional) integer/numeric value indicating column in \code{sample.mutations} which contain classification for SNV (Silent or not).
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Hugo_Symbol}{(optional) integer/numeric value indicating column in \code{sample.mutations} having gene names for reported SNVs.
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{Tumor_Sample_Barcode}{(optional) integer/numeric value indicating column in \code{sample.mutations} which have sample ids for SNVs.
Default is NULL value (in this case \code{sample.mutations} should already have this column)}

\item{CCF}{(optional) integer/numeric value indicating column in \code{sample.mutations} which have cancer cell fraction information for SNVs.
Default is NULL value (in this case \code{sample.mutations} should already have this column)}
}
\value{
a named numeric vector of probabilites that gene has nonsilent mutation (not caused by cancer).
}
\description{
\code{bcgr} function calculates  background probability that gene is mutated based on frequency od silent mutations.
}
\details{
With assumption of neutral selections, function estimates expected number of nonsilent mutations from observed number of silent mutations.
 Na (number of all possible nonsilent substitutions) and Ns (number of all possible silent substitutions) are taken from Lawrence paper.
 They are provdied in this package in file lawrence.RData.
 When expected number of nonsilent mutations for each gene is known, probability to get nonsilent mutation in each gene is calculated.
 This is based on
}
\examples{
\donttest{
# We first need CCF column
sample.genes.mutect <- CCF(sample.genes.mutect)
somatic.background <- bcgr(sample.genes.mutect, length.genes$Hugo_Symbol)
head(somatic.background)
}
}
\keyword{background}

